Target API 31 on Google variant
