+++ mobile/src/main/AndroidManifest.xml
@@ -52,10 +52,6 @@
     <uses-permission-sdk-23 android:name="android.permission.ACCESS_COARSE_LOCATION"/>
     <!-- Required since API 29 -->
     <uses-permission-sdk-23 android:name="android.permission.ACCESS_FINE_LOCATION"/>
-    <!-- Required since API 31, when targeting API 31 -->
-    <!--
-    <uses-permission-sdk-23 android:name="android.permission.BLUETOOTH_CONNECT"/>
-    -->
 
     <application
         android:name=".App"
